So, JS doesn't always calculate a perfect scroll on right side
This may be due to flex CSS stuff, or just temperamental browsers,
regardless, a small margin is applied to right side of tab to hide the
indicator when scrolled to the right. This way, if we are pretty close
the indicator goes away. Hopefully, I don't need to touch this again.